Move Quoted and Spliced extractors to TreeInfo
